Fixed infinite loop of ajax calls sending dispatches on render. Working on details modal bugs.
