+++ client/actionTypes.js
@@ -42,3 +42,4 @@ export const UPDATE_FORM_QUANTITY = 'UPDATE_FORM_QUANTITY';
 export const UPDATE_FORM_PURCHASE_PRICE = 'UPDATE_FORM_PURCHASE_PRICE';
 export const UPDATE_FORM_SELLER_SKU = 'UPDATE_FORM_SELLER_SKU';
 export const UPDATE_FORM_ASIN = 'UPDATE_FORM_ASIN';
+export const UPDATE_LAST_CHANGED_DETAIL_ID = 'UPDATE_LAST_CHANGED_DETAIL_ID';
