+++ client/store/initStore.js
@@ -5,9 +5,6 @@ import { routerMiddleware } from 'react-router-redux';
 import rootReducer from '../reducers';
 import initialState from './initialState';
 
-import { subscribeTo } from './subscribeTo';
-import smartDispatch from './dispatcher';
-import Backlog from './Backlog';
 
 const middleware = routerMiddleware(browserHistory);
 
@@ -17,11 +14,18 @@ const defaultStore = createStore(
   applyMiddleware(middleware)
 );
 
+export const store = defaultStore;
 
-// We want to extend the store to have subscribeTo, smartDispatch, register, and syncWithStore
-const backlog = new Backlog({
-  subscribeTo: subscribeTo.bind(defaultStore),
-  smartDispatch: smartDispatch.bind(defaultStore)
-});
 
-export const store = Object.assign({}, defaultStore, backlog);
+// // Old weird way
+// import { subscribeTo } from './subscribeTo';
+// import smartDispatch from './dispatcher';
+// import Backlog from './Backlog';
+
+// // We want to extend the store to have subscribeTo, smartDispatch, register, and syncWithStore
+// const backlog = new Backlog({
+//   subscribeTo: subscribeTo.bind(defaultStore),
+//   smartDispatch: smartDispatch.bind(defaultStore)
+// });
+
+// export const store = Object.assign({}, defaultStore, backlog);
